# NewPlaceBlazor/OpenApi/Api.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API",
        "version": "v1"
    },
    "paths": {
        "/api/Advertisement/search": {
            "get": {
                "tags": [
                    "Advertisement"
                ],
                "parameters": [
                    {
                        "name": "estateType",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdvertisementRepresentation"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdvertisementRepresentation"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdvertisementRepresentation"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Advertisement/details": {
            "get": {
                "tags": [
                    "Advertisement"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdvertisementDetailsRepresentation"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdvertisementDetailsRepresentation"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdvertisementDetailsRepresentation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Advertisement/create": {
            "post": {
                "tags": [
                    "Advertisement"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdvertisementDetailsRepresentation"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdvertisementDetailsRepresentation"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdvertisementDetailsRepresentation"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Authorization/login": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRepresentation"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRepresentation"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRepresentation"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserRepresentation"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserRepresentation"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserRepresentation"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Payment/donate": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DonateRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DonateRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/DonateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/Payment/top-donations": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "parameters": [
                    {
                        "name": "count",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PaymentDto"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PaymentDto"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PaymentDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Payment/status": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "parameters": [
                    {
                        "name": "paymentId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentStatusResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentStatusResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentStatusResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Recommendation/location": {
            "get": {
                "tags": [
                    "Recommendation"
                ],
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdvertisementDetailsRepresentation"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdvertisementDetailsRepresentation"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdvertisementDetailsRepresentation"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AdvertisementDetailsDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "apartmentType": {
                        "type": "string",
                        "nullable": true
                    },
                    "pricingType": {
                        "type": "string",
                        "nullable": true
                    },
                    "createDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "validTo": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "userName": {
                        "type": "string",
                        "nullable": true
                    },
                    "estateArea": {
                        "type": "number",
                        "format": "double"
                    },
                    "estateAddress": {
                        "type": "string",
                        "nullable": true
                    },
                    "estateCity": {
                        "type": "string",
                        "nullable": true
                    },
                    "utilities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StringDecimalNullableValueTuple"
                        },
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "provision": {
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "totalCost": {
                        "type": "string",
                        "nullable": true
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "additionalProperties": false
            },
            "AdvertisementDetailsRepresentation": {
                "type": "object",
                "properties": {
                    "resource": {
                        "$ref": "#/components/schemas/AdvertisementDetailsDto"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Link"
                        },
                        "nullable": true
                    },
                    "thumbnail": {
                        "$ref": "#/components/schemas/ImageRepresentation"
                    }
                },
                "additionalProperties": false
            },
            "AdvertisementDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "estateType": {
                        "type": "string",
                        "nullable": true
                    },
                    "pricingType": {
                        "type": "string",
                        "nullable": true
                    },
                    "userName": {
                        "type": "string",
                        "nullable": true
                    },
                    "estateArea": {
                        "type": "number",
                        "format": "double"
                    },
                    "estateAddress": {
                        "type": "string",
                        "nullable": true
                    },
                    "estateCity": {
                        "type": "string",
                        "nullable": true
                    },
                    "utilities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "provision": {
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "AdvertisementRepresentation": {
                "type": "object",
                "properties": {
                    "resource": {
                        "$ref": "#/components/schemas/AdvertisementDto"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Link"
                        },
                        "nullable": true
                    },
                    "thumbnail": {
                        "$ref": "#/components/schemas/ImageRepresentation"
                    }
                },
                "additionalProperties": false
            },
            "DonateRequest": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string",
                        "nullable": true
                    },
                    "amount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "currency": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "ImageRepresentation": {
                "type": "object",
                "properties": {
                    "resource": {
                        "type": "string",
                        "nullable": true
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Link"
                        },
                        "nullable": true
                    },
                    "mediaType": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "Link": {
                "type": "object",
                "properties": {
                    "rel": {
                        "type": "string",
                        "nullable": true
                    },
                    "href": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "method": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "PaymentDto": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "amount": {
                        "type": "number",
                        "format": "double"
                    },
                    "currency": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "PaymentStatusResponse": {
                "type": "object",
                "properties": {
                    "paymentId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": {}
            },
            "StringDecimalNullableValueTuple": {
                "type": "object",
                "additionalProperties": false
            },
            "UserDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "login": {
                        "type": "string",
                        "nullable": true
                    },
                    "passwordHash": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "UserRepresentation": {
                "type": "object",
                "properties": {
                    "resource": {
                        "$ref": "#/components/schemas/UserDto"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Link"
                        },
                        "nullable": true
                    },
                    "password": {
                        "type": "string",
                        "nullable": true
                    },
                    "token": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            }
        }
    }
}